ci: refactor hex handling
Move hex files definitions out of workflow files.
Remove duplicate constants definitions also.
